move norm, spectrogram and mfcc to preprocessor
